requirements.txt
matplotlib==3.0.3
RPi.GPIO==0.6.5
statistics==1.0.3.5
